fix errors

+++ src/RotationUtil.java
@@ -23,7 +23,7 @@ public class RotationUtil {
      * @param destAngle  Desired angle of target
      * @return direction enum
      */
-    static direction turnLeftOrRight(int robotAngle, int destAngle){
+    static direction turnLeftOrRight(double robotAngle, double destAngle){
         if (robotAngle > 180){
             robotAngle = robotAngle % 360;
             if (robotAngle > 180) robotAngle -= 180;
@@ -41,7 +41,7 @@ public class RotationUtil {
      * @param destAngle  Desired angle of target
      * @return direction enum
      */
-    static direction turnLeftOrRight180(int robotAngle, int destAngle){
+    static direction turnLeftOrRight180(double robotAngle, double destAngle){
         if (Math.abs (robotAngle - destAngle) <= 10) return direction.none;
 
         if (Math.abs(robotAngle - destAngle) < 180) {
